+++ js/sidebar.js
@@ -6,7 +6,14 @@ import {
   createErrorModal,
 } from "./utils";
 import { checkForAutosave, closeEditor, loadAutosave } from "./editor";
-import { getState, setState } from "./state";
+import {
+  addState,
+  checkState,
+  getState,
+  rmState,
+  sendState,
+  setState,
+} from "./state";
 
 const sidebar = document.querySelector("#sidebar");
 const folderStructure = document.querySelector("#folderStructure");
@@ -21,8 +28,39 @@ function createWrapper() {
   return wrapper;
 }
 
-// Creating search
+// Searching
 let key = "";
+// Gets search results from the server and applies them to the sidebar.
+async function doSearch(input) {
+  key = input.value;
+  if (!key) return;
+  const search = await fetch(`api/documents/search?key=${key}`, {
+    method: "GET",
+  });
+
+  if (search.ok) {
+    const results = await search.json();
+    searchIsOpen = true;
+
+    folderStructure.innerHTML = "";
+    folderStructure.appendChild(createSearch());
+
+    // Replacing Sidebar with Results
+    results.forEach((result) => {
+      const wrapper = createWrapper();
+      const file = createFile(
+        { name: result.name, isFolder: false },
+        result.folderPath + "/",
+      );
+      file.appendChild(createFileActions(result.name, result.folderPath + "/"));
+      wrapper.appendChild(setIcon("../assets/function/file.svg"));
+      wrapper.appendChild(file);
+      folderStructure.appendChild(wrapper);
+    });
+  }
+}
+
+// Creating search field & operations
 function createSearch() {
   const searchWrapper = document.createElement("div");
   searchWrapper.classList.add("wrapper");
@@ -42,34 +80,15 @@ function createSearch() {
   icon.src = "../assets/function/search.svg";
 
   // When the search icon is clicked, the search is activated.
-  icon.addEventListener("click", async () => {
-    key = input.value;
-    if (!key) return;
-    const search = await fetch(`api/documents/search?key=${key}`, {
-      method: "GET",
-    });
+  icon.addEventListener("click", () => {
+    doSearch(input);
+  });
 
-    if (search.ok) {
-      const results = await search.json();
-      searchIsOpen = true;
-
-      folderStructure.innerHTML = "";
-      folderStructure.appendChild(createSearch());
-
-      // Replacing Sidebar with Results
-      results.forEach((result) => {
-        const wrapper = createWrapper();
-        const file = createFile(
-          { name: result.name, isFolder: false },
-          result.folderPath + "/",
-        );
-        file.appendChild(
-          createFileActions(result.name, result.folderPath + "/"),
-        );
-        wrapper.appendChild(setIcon("../assets/function/file.svg"));
-        wrapper.appendChild(file);
-        folderStructure.appendChild(wrapper);
-      });
+  // Enter activates search.
+  input.addEventListener("keydown", (e) => {
+    if (e.key === "Enter") {
+      e.preventDefault();
+      doSearch(input);
     }
   });
 
@@ -139,11 +158,54 @@ function createFile(entry, previousEntry) {
   return file;
 }
 
+const expandedIcon = "../assets/function/expanded.svg";
+const collapsedIcon = "../assets/function/collapsed.svg";
+// Toggles expanded folders on the sidebar.
+function toggleExpanded(path) {
+  // If the folder is collapsed, it's expanded.
+  if (checkState("collapsedFolders", path)) {
+    rmState("collapsedFolders", path);
+    return expandedIcon;
+  } else {
+    addState("collapsedFolders", path);
+    return collapsedIcon;
+  }
+}
+
 // Creating an icon for the sidebar
-function setIcon(iconPath) {
+function setIcon(iconPath, path) {
+  // Current icon for collapsed / expanded. Local variable so that different folders don't get the icon that they're not supposed to.
+  let currentIcon;
+  // Gets the correct icon between expanded / collapsed.
+  if (checkState("collapsedFolders", path)) {
+    currentIcon = collapsedIcon;
+  } else {
+    currentIcon = expandedIcon;
+  }
+
   const icon = document.createElement("img");
   icon.classList.add("sidebarIcon");
   icon.src = iconPath;
+
+  if (iconPath !== "../assets/function/file.svg") {
+    // Hovering shows whether the folder is collapsed.
+    icon.addEventListener("mouseenter", () => {
+      icon.src = currentIcon;
+    });
+
+    // Reset icon
+    icon.addEventListener("mouseleave", () => {
+      icon.src = iconPath;
+    });
+
+    // Clicking toggles expanded / collapsed.
+    icon.addEventListener("click", () => {
+      currentIcon = toggleExpanded(path);
+      icon.src = currentIcon;
+      // Refreshes sidebar
+      buildSidebar();
+    });
+  }
   return icon;
 }
 
@@ -304,16 +366,30 @@ function renderEntries(entries, indentlevel, previousEntry) {
       folder.appendChild(createFolderActions(entries[i].name, previousEntry));
 
       if (indentlevel === 0) {
-        wrapper.appendChild(setIcon("../assets/function/notebook.svg"));
+        wrapper.appendChild(
+          setIcon(
+            "../assets/function/notebook.svg",
+            previousEntry + entries[i].name,
+          ),
+        );
       } else {
-        wrapper.appendChild(setIcon("../assets/function/folder.svg"));
+        wrapper.appendChild(
+          setIcon(
+            "../assets/function/folder.svg",
+            previousEntry + entries[i].name,
+          ),
+        );
       }
 
       wrapper.appendChild(folder);
       wrapper.style.marginLeft = 5 + indentlevel * 10 + "px";
       folderStructure.appendChild(wrapper);
 
-      if (entries[i].children.length > 0) {
+      if (
+        // Cancels rendering if the folder is collapsed.
+        entries[i].children.length > 0 &&
+        !checkState("collapsedFolders", previousEntry + entries[i].name)
+      ) {
         renderEntries(
           entries[i].children,
           indentlevel + 1,
@@ -325,7 +401,7 @@ function renderEntries(entries, indentlevel, previousEntry) {
       const wrapper = createWrapper();
       const file = createFile(entries[i], previousEntry);
       file.appendChild(createFileActions(entries[i].name, previousEntry));
-      wrapper.appendChild(setIcon("../assets/function/file.svg"));
+      wrapper.appendChild(setIcon("../assets/function/file.svg", ""));
       wrapper.appendChild(file);
       wrapper.style.marginLeft = 5 + indentlevel * 10 + "px";
       folderStructure.appendChild(wrapper);
@@ -450,3 +526,8 @@ rootButton.addEventListener("click", async (e) => {
     }
   });
 });
+
+// Updates master.json property "collapsedFolders" every 10s.
+setInterval(() => {
+  sendState("collapsedFolders");
+}, 5000);